gfx906: QSA-FN-1 - run the Qwen4Exp QSA path in fp16 (fixes the reported BF16 error; 4.4x kernel win)
The tester's `NotImplementedError: Qwen4Exp QSA currently requires BF16` is the
gfx906 bf16->fp16 auto-fallback (platforms/rocm.py:645, config/model.py:2294)
meeting QSA's bf16-only guards. This admits fp16 everywhere the QSA path stores
or reads a 2-byte float, via one shared pair of constants
(QSA_ACTIVATION_DTYPES / QSA_KV_CACHE_DTYPES in common/qsa_cache.py):

- amd/qsa.py: backend supported_dtypes/supported_kv_cache_dtypes, the Impl's
  KV-dtype guard, forward_qsa's Q/K/V guard (now also key_cache==query dtype),
  the attention activation guard (the reported error) and both cache guards.
- amd/indexer_qsa.py: the second copy of the activation guard; the raw and
  compressed index-key caches now take model_config.dtype.
- amd/ops/qsa.py: the sparse-attention dtype assert.
- amd/model.py x2, amd/mtp.py: HyperConnectionConfig(params_dtype=...) now takes
  the model dtype - bf16 HC weights under fp16 activations are not a coherent
  fp16 path. (Pull-in from QSA-FN-2.)
- common/qsa_cache.py: dtype-general (self.dtype / model dtype) so the shared
  NVIDIA path is untouched; _BF16_PER_INT64 -> _ELEMS_PER_INT64 (same value: 4 x
  2 B = 8 B = one int64, correct for both dtypes).

No kernel changed - fp16 rides the existing Triton kernels. The win is codegen:
gfx906 has no bf16 instruction, so bf16 tl.dot lowers to scalar v_fmac_f32 while
fp16 lowers to v_dot2_f32_f16. Measured on MI50 (launch-regime): sparse attention
26.5 ms fp16 vs 116.5 ms bf16 (4.39x, interleaved, rep-stable); per-row indexer
5426 vs 6928 us (1.28x).

Gates (all green, one boot):
- test_qsa_amd.py 9 -> 16 passed: the sparse-attention reference test is now
  parametrized bf16|fp16 (5 TP shapes), plus a new indexer-scoring reference test
  (qsa_mqa_paged, bf16+fp16) against a torch reference.
- test_qsa_reference.py 16 -> 19 passed: shared state-cache bind test parametrized
  bf16|fp16, plus a new test pinning the int64-MRoPE packing in both 2-byte dtypes.
- FA suite 104 passed; in-process PPL (Qwen3.8-27B-AWQ-INT4, fp16, 359 tokens)
  10.5472 with 0 top-20 misses, identical to the recorded value for this build;
  MoE 35B bench pp2048/tg256 4 samples 58.31/58.35/58.29/58.23 t/s (mean 58.30,
  mclk 1000) - parity with the 57.97/58.36 records.

Limits (recorded): no served request has exercised the fp16 path (the model needs
~60 GB of W4A16 weights), so the config-shape plumbing is uninstantiated and the
indexer's compress/store/selection kernels are untested in fp16. The end-to-end
gate is QSA-FN-3 (tiny-config harness).

Docs: DEVLOG-qwen38-flash-qsa.md (record), ROADMAP QSA-FN-1 SHIPPED / FN-7 gates
run / FN-2 reduced, CHANGELOG entry, RECON §1+§9 marked as the pre-edit inventory.

+++ vllm/models/qwen4_exp/amd/mtp.py
@@ -1,5 +1,6 @@
 # SPDX-License-Identifier: Apache-2.0
 # SPDX-FileCopyrightText: Copyright contributors to the vLLM project
+# SPDX-FileCopyrightText: Copyright Kevin Read <me@kevin-read.com>
 """Inference-only Qwen4Exp MTP (Multi-Token Predictor) model.
 
 The MTP draft model reuses the Qwen4Exp backbone (PLE/HC/MoE) but:
@@ -226,7 +227,7 @@ class Qwen4ExpMultiTokenPredictor(nn.Module):
         hc_config = HyperConnectionConfig(
             hc_count=config.hc_count,
             hidden_size=config.hidden_size,
-            params_dtype=torch.bfloat16,
+            params_dtype=model_config.dtype,
             hc_lowrank=config.hc_lowrank,
             rms_norm_eps=config.rms_norm_eps,
             hc_per_branch_norm=True,
